+++ Demo/Shared/Demo/DemoAppearance.swift
@@ -16,6 +16,8 @@ final class DemoAppearance {
     private init() {}
     
     static func apply() {
+        MenuListStyle.defaultStyle = .groupedWithInsets
+        
         #if os(iOS)
         if #available(iOS 13.0, *) {
             let navbar = UINavigationBar.appearance()